Update package.json and .gitignore
Moved "typescript" from devDependencies to dependencies in package.json. This is to ensure that Typescript is available in all environments, not just in development.

Also, updated .gitignore to include additional IDE-specific files. This is done to prevent unnecessary files that are specific to personal development setup from being tracked in version control.