Update to React 18
React 18 was integrated into the application for better performance and to take advantage of its latest features. Changes include changing "react-jsx" to "react" in tsconfig.json, adding a step for the installation of dependencies from package-lock.json and adjustments for the rendering of components in index.tsx. A React import was also added to Flow.tsx file, everything is working now.